# openapi.yaml
openapi: 3.0.3
info:
  title: E-Bookstore API
  description: Complete REST API for the E-Bookstore capstone application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName: { type: string, example: "Jane Doe" }
        email: { type: string, format: email, example: "jane@example.com" }
        password: { type: string, minLength: 8, example: "password123" }
        phone: { type: string, example: "+1-555-0100" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: "jane@example.com" }
        password: { type: string, example: "password123" }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        id: { type: integer, format: int64 }
        fullName: { type: string }
        email: { type: string }
        giftPoints: { type: integer }
    UserProfileResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        giftPoints: { type: integer }
        createdAt: { type: string, format: date-time }
    UpdateProfileRequest:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
...
    ProductSummaryResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        author: { type: string }
        price: { type: number, format: double }
        imageUrl: { type: string }
        rating: { type: number, format: double }
        categoryName: { type: string }
        brandName: { type: string }
        stockQuantity: { type: integer }
...
    AddToCartRequest:
      type: object
      required: [productId]
      properties:
        productId: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1, default: 1 }
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1 }
    CartItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        product: { $ref: '#/components/schemas/ProductSummaryResponse' }
        quantity: { type: integer }
        itemTotal: { type: number, format: double }
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItemResponse' }
        subtotal: { type: number, format: double }
        shipping: { type: number, format: double }
        total: { type: number, format: double }
        itemCount: { type: integer }
...
    AddressRequest:
      type: object
      properties:
        label: { type: string }
        street: { type: string }
        city: { type: string }
        state: { type: string }
        zipCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
    AddressResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        label: { type: string }
        street: { type: string }
        city: { type: string }
        state: { type: string }
        zipCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      summary: Register a new user
      description: Creates a new user account and returns a JWT token for immediate login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            example:
              fullName: "Jane Doe"
              email: "jane@example.com"
              password: "password123"
              phone: "+1-555-0100"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Validation error or email already taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/login:
    post:
      tags: [Authentication]
      operationId: loginUser
      summary: Authenticate user
      description: Validates credentials and returns a JWT token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              email: "jane@example.com"
              password: "password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/me:
    get:
      tags: [Authentication]
      operationId: getProfile
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfileResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      tags: [Authentication]
      operationId: updateProfile
      summary: Update current user profile
      description: Updates full name and/or phone number for the authenticated user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfileResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/cart:
    get:
      tags: [Cart]
      operationId: getCart
      summary: Get current user's cart
      description: Returns the shopping cart for the authenticated user including subtotal, shipping and total.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cart contents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Cart]
      operationId: clearCart
      summary: Clear the cart
      description: Removes all items from the authenticated user's cart.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Cart cleared
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/cart/items:
    post:
      tags: [Cart]
      operationId: addToCart
      summary: Add item to cart
      description: Adds a product to the cart. If the product already exists, quantity is incremented.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddToCartRequest' }
            example: { productId: 1, quantity: 2 }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartResponse' }
        '400':
          description: Out of stock or invalid quantity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Product not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/cart/items/{id}:
    put:
      tags: [Cart]
      operationId: updateCartItem
      summary: Update cart item quantity
      description: Sets a new quantity for a specific cart item.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCartItemRequest' }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartResponse' }
        '400':
          description: Invalid quantity or insufficient stock
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Cart item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Cart]
      operationId: removeCartItem
      summary: Remove item from cart
      description: Removes a specific item from the cart.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Cart item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/addresses:
    get:
      tags: [Addresses]
      operationId: getAddresses
      summary: Get user addresses
      description: Returns all saved delivery addresses for the authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AddressResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Addresses]
      operationId: addAddress
      summary: Add a new address
      description: Creates a new delivery address for the authenticated user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressRequest' }
            example:
              label: "Home"
              street: "123 Main St"
              city: "New York"
              state: "NY"
              zipCode: "10001"
              country: "USA"
              isDefault: true
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddressResponse' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/addresses/{id}:
    put:
      tags: [Addresses]
      operationId: updateAddress
      summary: Update an address
      description: Updates an existing delivery address belonging to the authenticated user.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressRequest' }
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddressResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Address does not belong to user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Address not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Addresses]
      operationId: deleteAddress
      summary: Delete an address
      description: Deletes a delivery address belonging to the authenticated user.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Address deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Address does not belong to user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Address not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/addresses/{id}/default:
    put:
      tags: [Addresses]
      operationId: setDefaultAddress
      summary: Set address as default
      description: Marks the specified address as the default delivery address, clearing the flag on all others.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Default address updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddressResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Address does not belong to user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Address not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...